refactor(US15): add folder services in admin/shared and replace all service.ts in there
