+++ src/app/admin/shared/Services/auth.guard.service.ts
@@ -3,7 +3,7 @@ import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
 
 import {AuthService} from './auth.service';
-import { ROUTE_CONFIGS } from '../../utils/constants/route.consts';
+import { ROUTE_CONFIGS } from '../../../utils/constants/route.consts';
 
 @Injectable()
 export class AuthGuardService implements CanActivate {
