+++ src/app/admin/shared/components/admin-layout/admin-layout.component.html
@@ -1,3 +1,5 @@
+<app-alert></app-alert>
+
 <nav class="navbar bg-primary">
   <h1>
     <a routerLink="/">Angular Admin</a>
@@ -13,7 +15,7 @@
     </li>
     <li routerLinkActive="active">
       <a [routerLink]="[
-      routeConfig.adminPage.fullpath,
+       routeConfig.adminPage.fullpath,
        routeConfig.adminCreate.path
        ]">
         Создать
